Added new effect for objects getting damaged, fixed particle lifetime, fixed some ui elements not drawing fully. Also seem to have accidently ran a code formatter on gameobject.c oops

+++ particle.c
@@ -49,7 +49,7 @@ void AddParticleWithRandomProperties(float x, float y, Color color, float lifeti
     AddParticle(x,y,RandRange(lifetimeMin,lifetimeMax),RandRange(speedMin,speedMax),RandRange(minAngle,maxAngle),color);
 }
 
-void AddParticle(float x, float y, short lifetime, float speed, float angle, Color colour)
+void AddParticle(float x, float y, float lifetime, float speed, float angle, Color colour)
 {
     particle_x[PARTICLES_TOP] = x;
     particle_y[PARTICLES_TOP] = y;
