+++ particle.h
@@ -30,7 +30,7 @@ void ClearParticles();
 void UpdateParticles(float dt);
 void DrawParticles();
 void InitParticles();
-void AddParticle(float x, float y, short lifetime, float speed, float angle, Color colour); 
+void AddParticle(float x, float y, float lifetime, float speed, float angle, Color colour); 
 void AddMouseRandomParticles(MouseState mouseState, int numParticles);
 void RandParticleAroundEdgeOfCircle(float cx, float cy, float r, float lifetime, float maxspeed, Color col);
 void AddParticleWithRandomProperties(float x, float y, Color color, float lifetimeMin, float lifetimeMax, float speedMin, float speedMax, float minAngle, float maxAngle);
